requirements.txt
aerich
aiogram==3.0.0
aiohttp
aiosqlite
aioschedule
cachetools
coloredlogs
toml
tortoise-orm
gitpython
asyncpg
Faker